+++ CHANGELOG.md
@@ -17,6 +17,7 @@
     - Added PPT Template.
     - Modified PPT to split nodes in chunks of 50 nodes per slide (2 columns of 50 nodes).
     - Persist Allowed SSB and ARFCN lists between different executions.
+    - Improvements on Configuration Audit.
     
   - #### 🐛 Bug fixes:
     - Minor bug fixing.